Working CRC32 checks on all non-map pages

+++ speeduino/comms.h
@@ -44,7 +44,7 @@ const char pageTitles[] PROGMEM //This is being stored in the avr flash instead 
 
 void command();//This is the heart of the Command Line Interpeter.  All that needed to be done was to make it human readable.
 void sendValues(uint16_t, uint16_t,byte, byte);
-void receiveValue(int, byte);
+void receiveValue(uint16_t, byte);
 void saveConfig();
 void sendPage(bool);
 void receiveCalibration(byte);
